chore: create common package for duprecated definitions
