+++ cmd/dump/dump.go
@@ -7,6 +7,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	root "github.com/koh-sh/codebuild-multirunner/cmd"
+	"github.com/koh-sh/codebuild-multirunner/common"
 	"github.com/spf13/cobra"
 )
 
@@ -15,7 +16,7 @@ var dumpCmd = &cobra.Command{
 	Use:   "dump",
 	Short: "dump config for running CodeBuild projects",
 	Run: func(cmd *cobra.Command, args []string) {
-		bc := root.ReadConfigFile(root.Configfile)
+		bc := common.ReadConfigFile(root.Configfile)
 		fmt.Println(dumpConfig(bc))
 	},
 }
@@ -25,7 +26,7 @@ func init() {
 }
 
 // dump read config with environment variables inserted
-func dumpConfig(bc root.BuildConfig) string {
+func dumpConfig(bc common.BuildConfig) string {
 	d, err := yaml.Marshal(&bc)
 	if err != nil {
 		log.Fatal(err)
