+++ cmd/run/run_test.go
@@ -7,13 +7,12 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/service/codebuild"
 	"github.com/aws/aws-sdk-go-v2/service/codebuild/types"
-	"github.com/fatih/color"
-	root "github.com/koh-sh/codebuild-multirunner/cmd"
+	"github.com/koh-sh/codebuild-multirunner/common"
 )
 
 func Test_convertBuildConfigToStartBuildInput(t *testing.T) {
 	type args struct {
-		build root.Build
+		build common.Build
 	}
 	tests := []struct {
 		name string
@@ -21,7 +20,7 @@ func Test_convertBuildConfigToStartBuildInput(t *testing.T) {
 		want codebuild.StartBuildInput
 	}{
 		{name: "basic",
-			args: args{root.Build{}},
+			args: args{common.Build{}},
 			want: codebuild.StartBuildInput{},
 		},
 	}
@@ -34,91 +33,11 @@ func Test_convertBuildConfigToStartBuildInput(t *testing.T) {
 	}
 }
 
-func Test_coloredString(t *testing.T) {
-	type args struct {
-		status string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{name: "SUCCEEDED",
-			args: args{status: "SUCCEEDED"},
-			want: color.GreenString("SUCCEEDED")},
-		{name: "IN_PROGRESS",
-			args: args{status: "IN_PROGRESS"},
-			want: color.BlueString("IN_PROGRESS")},
-		{name: "FAILED",
-			args: args{status: "FAILED"},
-			want: color.RedString("FAILED")},
-		{name: "TIMED_OUT",
-			args: args{status: "TIMED_OUT"},
-			want: color.RedString("TIMED_OUT")},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := coloredString(tt.args.status); got != tt.want {
-				t.Errorf("coloredString() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_buildStatusCheck(t *testing.T) {
-	var id1 = "project:12345678"
-	var id2 = "project2:87654321"
-	var ids = []string{id1, id2}
-
-	type args struct {
-		client func(t *testing.T) root.CodeBuildAPI
-		ids    []string
-	}
-	tests := []struct {
-		name  string
-		args  args
-		want  []string
-		want2 bool
-	}{
-		{name: "all builds ended",
-			args:  args{client: root.ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "SUCCEEDED", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: false,
-		},
-		{name: "one builds in progress",
-			args:  args{client: root.ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "IN_PROGRESS", Id: &id2}}), ids: ids},
-			want:  []string{id2},
-			want2: false,
-		},
-		{name: "one of builds failed",
-			args:  args{client: root.ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "FAILED", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: true,
-		},
-		{name: "one of builds timeout",
-			args:  args{client: root.ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "TIMED_OUT", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got2 := buildStatusCheck(tt.args.client(t), tt.args.ids)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got, tt.want)
-			}
-			if !reflect.DeepEqual(got2, tt.want2) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got2, tt.want2)
-			}
-		})
-	}
-}
-
 func Test_runCodeBuild(t *testing.T) {
 	var project = "project"
 	var id = "project:12345"
 	type args struct {
-		client func(t *testing.T) root.CodeBuildAPI
+		client func(t *testing.T) common.CodeBuildAPI
 		input  codebuild.StartBuildInput
 	}
 	tests := []struct {
@@ -128,12 +47,12 @@ func Test_runCodeBuild(t *testing.T) {
 		wantErr bool
 	}{
 		{name: "success to start",
-			args:    args{client: root.ReturnStartBuildMockAPI(&types.Build{Id: &id}, nil), input: codebuild.StartBuildInput{ProjectName: &project}},
+			args:    args{client: common.ReturnStartBuildMockAPI(&types.Build{Id: &id}, nil), input: codebuild.StartBuildInput{ProjectName: &project}},
 			want:    id,
 			wantErr: false,
 		},
 		{name: "fail to start",
-			args:    args{client: root.ReturnStartBuildMockAPI(&types.Build{Id: &id}, errors.New("fail to run")), input: codebuild.StartBuildInput{ProjectName: &project}},
+			args:    args{client: common.ReturnStartBuildMockAPI(&types.Build{Id: &id}, errors.New("fail to run")), input: codebuild.StartBuildInput{ProjectName: &project}},
 			want:    "",
 			wantErr: true,
 		},
